// src/common/constants/exam-template.enum.ts
export enum TemplateMode {
  PRACTICE = 'practice',
  MOCK_TEST = 'mock_test',
  OFFICIAL_EXAM = 'official_exam',
}

export enum TemplateStatus {
  DRAFT = 'draft',
  PUBLISHED = 'published',
  ARCHIVED = 'archived',
}

export enum TemplateItemMode {
  MANUAL = 'manual',
  RULE_BASED = 'rule_based',
}